Refactor HelpController to an action.
Uses the normal layout (lightbox or full page) for help pages. The separate layout/help.phtml layout and css/help.css have been removed.
