--- themes/root/css/help.css
@@ -1,40 +1,0 @@
-body {
-  font-family: Arial, Helvetica;
-  font-size: 10pt;
-}
-
-ul.HelpMenu {
-  float: left;
-  width: 150px;
-  background-color: #FFFFCC;
-  margin:0 3px 0 0;
-  border:1px solid #DD0;
-}
-
-dl.Content {
-  border-left: solid 1px #999999;
-  padding-left: 10px;
-  height: 430px;
-  overflow: auto;
-}
-
-dl dt {
-  font-weight: bold;
-}
-
-dl dd {
-  padding: 0px;
-  margin: 2px 1px 2px 10px;
-}
-
-pre {
-  padding: 10px;
-  margin: 10px;
-  background-color: #EEEEEE;
-  border: dashed 1px #999999;
-  width: 300px;
-}
-
-.warning {
-  color: red;
-}
