--- themes/root/templates/layout/help.phtml
@@ -1,15 +1,0 @@
-<?=$this->doctype()?>
-<html lang="<?=$this->layout()->userLang?>"<?php if ($this->layout()->rtl): ?> dir="rtl"<?php endif; ?>>
-  <head>
-    <?php
-      $this->setupThemeResources();
-      $this->assetManager()->clearScriptList()->clearStyleList()->appendStyleLink('help.css');
-    ?>
-    <?=$this->headMeta()?>
-    <?=$this->headTitle()?>
-    <?=$this->assetManager()->outputHeaderAssets()?>
-  </head>
-  <body>
-    <?=$this->layout()->content?>
-  </body>
-</html>
